backend/requirements.txt
anthropic==0.49.0
fastapi==0.115.6
uvicorn[standard]==0.34.0
python-dotenv==1.0.1
pydantic==2.10.4
httpx==0.28.1
aiohttp==3.11.11
python-multipart==0.0.20
